w3c-reports/12092014/efl/runned-tests/html/semantics/embedded-content/media-elements/track/track-element/cors/046-expected.txt

CONSOLE MESSAGE: Cross-origin text track load denied by Cross-Origin Resource Sharing policy.
Actual	Expected
{
  "event": "error",
  "requests": [
    {
      "cors": "http://localhost:8800",
      "cookie": "yes"
    },
    {
      "cors": "http://localhost:8800",
      "cookie": "yes"
    }
  ]
}
{
  "event": "load",
  "requests": [
    {
      "cors": "no",
      "cookie": "yes"
    },
    {
      "cors": "no",
      "cookie": "yes"
    },
    {
      "cors": "http://localhost:8800",
      "cookie": "yes"
    }
  ]
}
Test id: 448482fa-e0af-46ad-966f-7718f686bbcf



FAIL track CORS: Use Credentials, same-origin, no headers, redirects to same-origin, no headers, redirects to not same-origin, with headers assert_object_equals: property "event" expected object "[object Object]" got object "[object Object]"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
check_equal@http://localhost:8800/resources/testharness.js:378:28
assert_object_equals@http://localhost:8800/resources/testharness.js:390:21
checkData@http://localhost:8800/html/semantics/embedded-content/media-elements/track/track-element/cors/support/common.js?pipe=sub:128:29
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:827:40)
